services/document-parser/requirements.txt
# FastAPI + server
fastapi==0.104.1
uvicorn==0.24.0

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# PDF parsing
pdfplumber==0.10.3
PyMuPDF==1.23.8

# OCR for scanned documents
pytesseract==0.3.10

# Document parsing
python-docx==1.0.1
openpyxl==3.1.0

# Formula parsing and units
sympy==1.12
pint==0.23.0

# Configuration
python-dotenv==1.0.0

# HTTP client
httpx==0.25.0
